Fixed precision loss in BottomDialogChatSettings due to float to double conversion
(cherry picked from commit e7128c2cec45ce9a27028df05627fec5f04c26ed)

+++ app/src/main/java/com/pandacorp/togetheraichat/domain/repository/TogetherRepository.kt
@@ -7,9 +7,9 @@ import kotlinx.coroutines.flow.Flow
 interface TogetherRepository {
     fun getResponse(
         messages: MutableList<MessageItem>,
-        temperature: Double,
+        temperature: Float,
         maxTokens: Int? = null,
-        frequencyPenalty: Double? = null,
-        topP: Double? = null
+        frequencyPenalty: Float? = null,
+        topP: Float? = null
     ): Flow<ChatCompletionChunk>
 }
